+++ src/control_plane/workload/mod.rs
@@ -1,0 +1,10 @@
+#[cfg(test)]
+mod tests;
+
+mod application_container_plan;
+mod application_container_plan_options;
+mod workload_plan_error;
+
+pub(crate) use application_container_plan::ApplicationContainerPlan;
+pub(crate) use application_container_plan_options::ApplicationContainerPlanOptions;
+pub(crate) use workload_plan_error::WorkloadPlanError;